requirements.txt
tensorflow-gpu 1.15.1
numpy 1.18.5
scipy 1.7.0
pandas 1.3.0
python 3.8.10